SG2042: Fixbug, exception occur when booting
In resize_dtb, we are trying to realloc memory from where the
device tree loaded. But, the memory where device tree is loaded
is not a allocated memory. Instead it is a fixed magic number.

The memory region which holds the device tree always large
enough for resizing a device tree. So don't realloc, just
open the device tree with the new size in place.

Signed-off-by: Chao Wei <chao.wei@sophgo.com>